+++ build.gradle
@@ -169,6 +169,7 @@ buildscript {
     classpath "cz.malohlava:visteg:1.0.3"                                      // Enable generating Gradle task dependencies as ".dot" files
     classpath "com.github.jengelman.gradle.plugins:shadow:2.0.1"               // Enable shading Java dependencies
     classpath "ca.coglinc:javacc-gradle-plugin:2.4.0"                          // Enable the JavaCC parser generator
+    classpath "org.junit.platform:junit-platform-gradle-plugin:1.0.2"
   }
 }
