+++ pom.xml
@@ -1791,6 +1791,7 @@
             <redirectTestOutputToFile>true</redirectTestOutputToFile>
             <argLine>${beamSurefireArgline}</argLine>
           </configuration>
+          <!-- let children do something else and it is the default anyway
           <dependencies>
             <dependency>
               <groupId>org.apache.maven.surefire</groupId>
@@ -1798,6 +1799,7 @@
               <version>${surefire-plugin.version}</version>
             </dependency>
           </dependencies>
+          -->
         </plugin>
 
         <plugin>
