+++ sdks/java/core/src/test/java/org/apache/beam/sdk/testing/SerializableMatchersTest.java
@@ -50,114 +50,114 @@ import org.junit.runners.JUnit4;
  */
 @RunWith(JUnit4.class)
 public class SerializableMatchersTest implements Serializable {
-  @Rule
-  public transient ExpectedException thrown = ExpectedException.none();
+    @Rule
+    public transient ExpectedException thrown = ExpectedException.none();
 
-  @Test
-  public void testAnythingSerializable() throws Exception {
-    SerializableUtils.ensureSerializable(anything());
-  }
-
-  @Test
-  public void testAllOfSerializable() throws Exception {
-    SerializableUtils.ensureSerializable(allOf(anything()));
-  }
-
-  @Test
-  public void testContainsInAnyOrderSerializable() throws Exception {
-    assertThat(ImmutableList.of(2, 1, 3),
-        SerializableUtils.ensureSerializable(containsInAnyOrder(1, 2, 3)));
-  }
-
-  @Test
-  public void testContainsInAnyOrderNotSerializable() throws Exception {
-    assertThat(
-        ImmutableList.of(new NotSerializableClass()),
-        SerializableUtils.ensureSerializable(containsInAnyOrder(
-            new NotSerializableClassCoder(),
-            new NotSerializableClass())));
-  }
-
-  @Test
-  public void testKvKeyMatcherSerializable() throws Exception {
-    assertThat(
-        KV.of("hello", 42L),
-        SerializableUtils.ensureSerializable(kvWithKey("hello")));
-  }
-
-  @Test
-  public void testKvMatcherBasicSuccess() throws Exception {
-    assertThat(
-        KV.of(1, 2),
-        SerializableMatchers.<Integer, Integer>kv(anything(), anything()));
-  }
-
-  @Test
-  public void testKvMatcherKeyFailure() throws Exception {
-    try {
-      assertThat(
-          KV.of(1, 2),
-          SerializableMatchers.<Integer, Integer>kv(not(anything()), anything()));
-    } catch (AssertionError exc) {
-      assertThat(exc.getMessage(), Matchers.containsString("key did not match"));
-      return;
-    }
-    fail("Should have failed");
-  }
-
-  @Test
-  public void testKvMatcherValueFailure() throws Exception {
-    try {
-      assertThat(
-          KV.of(1, 2),
-          SerializableMatchers.<Integer, Integer>kv(anything(), not(anything())));
-    } catch (AssertionError exc) {
-      assertThat(exc.getMessage(), Matchers.containsString("value did not match"));
-      return;
-    }
-    fail("Should have failed");
-  }
-
-  @Test
-  public void testKvMatcherGBKLikeSuccess() throws Exception {
-    assertThat(
-        KV.of("key", ImmutableList.of(1, 2, 3)),
-        SerializableMatchers.<Object, Iterable<Integer>>kv(
-            anything(), containsInAnyOrder(3, 2, 1)));
-  }
-
-  @Test
-  public void testKvMatcherGBKLikeFailure() throws Exception {
-    try {
-      assertThat(
-          KV.of("key", ImmutableList.of(1, 2, 3)),
-          SerializableMatchers.<String, Iterable<Integer>>kv(
-              anything(), containsInAnyOrder(1, 2, 3, 4)));
-    } catch (AssertionError exc) {
-      assertThat(exc.getMessage(), Matchers.containsString("value did not match"));
-      return;
-    }
-    fail("Should have failed.");
-  }
-
-  private static class NotSerializableClass {
-    @Override public boolean equals(Object other) {
-      return other instanceof NotSerializableClass;
+    @Test
+    public void testAnythingSerializable() throws Exception {
+        SerializableUtils.ensureSerializable(anything());
     }
 
-    @Override public int hashCode() {
-      return 0;
-    }
-  }
-
-  private static class NotSerializableClassCoder extends AtomicCoder<NotSerializableClass> {
-    @Override
-    public void encode(NotSerializableClass value, OutputStream outStream) {
+    @Test
+    public void testAllOfSerializable() throws Exception {
+        SerializableUtils.ensureSerializable(allOf(anything()));
     }
 
-    @Override
-    public NotSerializableClass decode(InputStream inStream) {
-      return new NotSerializableClass();
+    @Test
+    public void testContainsInAnyOrderSerializable() throws Exception {
+        assertThat(ImmutableList.of(2, 1, 3),
+                SerializableUtils.ensureSerializable(containsInAnyOrder(1, 2, 3)));
     }
-  }
+
+    @Test
+    public void testContainsInAnyOrderNotSerializable() throws Exception {
+        assertThat(
+                ImmutableList.of(new NotSerializableClass()),
+                SerializableUtils.ensureSerializable(containsInAnyOrder(
+                        new NotSerializableClassCoder(),
+                        new NotSerializableClass())));
+    }
+
+    @Test
+    public void testKvKeyMatcherSerializable() throws Exception {
+        assertThat(
+                KV.of("hello", 42L),
+                SerializableUtils.ensureSerializable(kvWithKey("hello")));
+    }
+
+    @Test
+    public void testKvMatcherBasicSuccess() throws Exception {
+        assertThat(
+                KV.of(1, 2),
+                SerializableMatchers.<Integer, Integer>kv(anything(), anything()));
+    }
+
+    @Test
+    public void testKvMatcherKeyFailure() throws Exception {
+        try {
+            assertThat(
+                    KV.of(1, 2),
+                    SerializableMatchers.<Integer, Integer>kv(not(anything()), anything()));
+        } catch (AssertionError exc) {
+            assertThat(exc.getMessage(), Matchers.containsString("key did not match"));
+            return;
+        }
+        fail("Should have failed");
+    }
+
+    @Test
+    public void testKvMatcherValueFailure() throws Exception {
+        try {
+            assertThat(
+                    KV.of(1, 2),
+                    SerializableMatchers.<Integer, Integer>kv(anything(), not(anything())));
+        } catch (AssertionError exc) {
+            assertThat(exc.getMessage(), Matchers.containsString("value did not match"));
+            return;
+        }
+        fail("Should have failed");
+    }
+
+    @Test
+    public void testKvMatcherGBKLikeSuccess() throws Exception {
+        assertThat(
+                KV.of("key", ImmutableList.of(1, 2, 3)),
+                SerializableMatchers.<Object, Iterable<Integer>>kv(
+                        anything(), containsInAnyOrder(3, 2, 1)));
+    }
+
+    @Test
+    public void testKvMatcherGBKLikeFailure() throws Exception {
+        try {
+            assertThat(
+                    KV.of("key", ImmutableList.of(1, 2, 3)),
+                    SerializableMatchers.<String, Iterable<Integer>>kv(
+                            anything(), containsInAnyOrder(1, 2, 3, 4)));
+        } catch (AssertionError exc) {
+            assertThat(exc.getMessage(), Matchers.containsString("value did not match"));
+            return;
+        }
+        fail("Should have failed.");
+    }
+
+    private static class NotSerializableClass {
+        @Override public boolean equals(Object other) {
+            return other instanceof NotSerializableClass;
+        }
+
+        @Override public int hashCode() {
+            return 0;
+        }
+    }
+
+    private static class NotSerializableClassCoder extends AtomicCoder<NotSerializableClass> {
+        @Override
+        public void encode(NotSerializableClass value, OutputStream outStream) {
+        }
+
+        @Override
+        public NotSerializableClass decode(InputStream inStream) {
+            return new NotSerializableClass();
+        }
+    }
 }
